irkutskoiltest/requirements.txt
Django==4.1
psycopg2-binary
djangorestframework==3.14.0
pandas==2.1.4
openpyxl==3.1.2
numpy
